requirements.txt
numpy==1.24.2
opencv-python==4.7.0.68
Pillow==9.4.0
pyfiglet==0.7
